src/models/db_schemas/data_chunk.py
from pydantic import BaseModel, Field, validator
from typing import Optional
from bson.objected import ObjectId

class DataChunk(BaseModel):
    _id: Optional[ObjectId]
    chunk_text: str = Field(..., min_length=1)
    chunk_metadata: dict
    chunk_order: int = Field(..., gt=0)
    chunk_project_id: ObjectId

    class Config:
        arbirtraty_types_allowed = True