src/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
python-dotenv==1.2.1
pydantic-settings==2.12.0
aiofiles==25.1.0
langchain==1.2.0
PyMuPDF==1.26.7
langchain-community==0.4.1
motor==3.7.1

